requirements.txt
line-bot-sdk==3.17.0
flask==3.0.0
python-dotenv==1.1.1